Add negative query type narrowing tests
